test: move e2e HTTP header tests to http_header_end2end_test.go
